+++ src/C++/ThreadedSocketConnection.cpp
@@ -69,7 +69,7 @@ bool ThreadedSocketConnection::send( const std::string& msg )
   int totalSent = 0;
   while(totalSent < (int)msg.length())
   {
-    int sent = socket_send( m_socket, msg.c_str() + totalSent, msg.length() );
+    ssize_t sent = socket_send( m_socket, msg.c_str() + totalSent, msg.length() );
     if(sent < 0) return false;
     totalSent += sent;
   }
@@ -101,7 +101,7 @@ bool ThreadedSocketConnection::read()
     if( result > 0 ) // Something to read
     {
       // We can read without blocking
-      int size = recv( m_socket, m_buffer, sizeof(m_buffer), 0 );
+      ssize_t size = recv( m_socket, m_buffer, sizeof(m_buffer), 0 );
       if ( size <= 0 ) { throw SocketRecvFailed( size ); }
       m_parser.addToStream( m_buffer, size );
     }
